[#2] Common/Data: limit paragraphs to PARAS_LIM (100 default)
Pages which have a large number of misses -- for example, if they are
not actually human text, or are in the wrong language -- result in an
excessive number of paragraphs. These take a long time to run through
Hunspell, and have dubious value, since most misses are likely
false-positives or indicative or some other problem. Limit the number of
paragraphs to `PARAS_LIM`, setting at `100 paragraphs` by default.